:globe_with_meridians: Use system locale for month names
POSIX-compliant systems provide an API for retrieving the system locale
from the system, as well as retrieving month names using this locale. This
commit makes the change to utilize this API, instead of trying to figure it
out independently and maintain a list of month names for a series of
languages.